+++ application/tests/_ci_phpunit_test/ChangeLog.md
@@ -2,7 +2,7 @@
 
 ## v0.17.4 (Not Released)
 
-* Fix bug that MY_Config does not work. See [#327](https://github.com/kenjis/ci-phpunit-test/pull/327).
+* Fix bug that config files are not loaded with HMVC. See [#327](https://github.com/kenjis/ci-phpunit-test/pull/327), [#328](https://github.com/kenjis/ci-phpunit-test/pull/328).
 
 ## v0.17.3 (2020/02/05)